refactor: dto 검증 어노테이션 추가 및 패키지 위치 이동
