+++ src/main/java/cart/dto/request/CartItemRequest.java
@@ -1,6 +1,10 @@
 package cart.dto.request;
 
+import javax.validation.constraints.NotNull;
+
 public class CartItemRequest {
+
+    @NotNull
     private Long productId;
 
     public CartItemRequest() {
