+++ src/test/java/cart/integration/CartItemIntegrationTest.java
@@ -7,8 +7,8 @@ import cart.domain.member.Member;
 import cart.domain.member.Password;
 import cart.dto.request.CartItemQuantityUpdateRequest;
 import cart.dto.request.CartItemRequest;
-import cart.dto.request.CartItemResponse;
 import cart.dto.request.ProductRequest;
+import cart.dto.response.CartItemResponse;
 import cart.repository.MemberRepository;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
