+++ src/test/java/cart/integration/OrderIntegrationTest.java
@@ -7,8 +7,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import cart.domain.member.Email;
 import cart.domain.member.Member;
 import cart.domain.member.Password;
-import cart.dto.request.CartItemResponse;
 import cart.dto.request.OrderRequest;
+import cart.dto.response.CartItemResponse;
 import cart.dto.response.OrderItemResponse;
 import cart.dto.response.OrderResponse;
 import io.restassured.response.ExtractableResponse;
